requirements.txt
customtkinter>=5.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pyinstaller>=5.0.0
